Get / Update consecutive days nb on firestore

+++ app/src/main/java/com/example/emafoods/feature/game/data/datasource/GameDataSource.kt
@@ -16,7 +16,7 @@ interface GameDataSource {
     suspend fun storeXpToUnspent(xp: Long)
     suspend fun unspentUserXP(): Long
     suspend fun resetUnspentXp()
-    suspend fun consecutiveDaysAppOpened(): Int
+    suspend fun consecutiveDaysAppOpened(): Long
     suspend fun updateConsecutiveDaysAppOpened()
     suspend fun resetConsecutiveDaysAppOpened()
     suspend fun addRewardToUserAcceptedRecipe(rewardedUserUid: String)
