+++ app/src/main/java/com/example/emafoods/feature/profile/presentation/ProfileViewModel.kt
@@ -67,7 +67,7 @@ class ProfileViewModel @Inject constructor(
             val streaks = consecutiveDaysAppOpenedUseCase.execute()
             _state.update {
                 it.copy(
-                    streaks = streaks
+                    streaks = streaks.toInt()
                 )
             }
         }
